+++ src/dialogs/versiondialog.cpp
@@ -5,10 +5,10 @@
 #include <QScriptValueIterator>
 #include <QPushButton>
 
-VersionDialog::VersionDialog(QScriptValue versions, MainWindow *mainWindow, QWidget *parent) :
+VersionDialog::VersionDialog(QScriptValue versions, MainWindow *mainWindow,
+                             QWidget *parent) :
         MasterDialog(parent),
-    ui(new Ui::VersionDialog)
-{
+        ui(new Ui::VersionDialog) {
     this->mainWindow = mainWindow;
     ui->setupUi(this);
 
@@ -17,20 +17,27 @@ VersionDialog::VersionDialog(QScriptValue versions, MainWindow *mainWindow, QWid
     QPushButton *button;
     ui->buttonBox->clear();
 
-    button = new QPushButton( tr( "&Restore selected version" ) );
-    button->setProperty( "ActionRole", Restore );
-    button->setDefault( false );
-    button->setIcon( QIcon( ":/images/breeze/edit-download.svg" ) );
-    ui->buttonBox->addButton( button, QDialogButtonBox::ActionRole );
+    button = new QPushButton(tr("&Restore selected version"));
+    button->setProperty("ActionRole", Restore);
+    button->setDefault(false);
+    button->setIcon(
+            QIcon::fromTheme(
+                    "edit-download",
+                    QIcon(":/icons/breeze-qownnotes/16x16/edit-download.svg")));
+    ui->buttonBox->addButton(button, QDialogButtonBox::ActionRole);
 
-    button = new QPushButton( tr( "&Cancel" ) );
-    button->setProperty( "ActionRole", Cancel );
-    button->setIcon( QIcon( ":/images/breeze/dialog-cancel.svg" ) );
-    button->setDefault( true );
-    ui->buttonBox->addButton( button, QDialogButtonBox::ActionRole );
+    button = new QPushButton(tr("&Cancel"));
+    button->setProperty("ActionRole", Cancel);
+    button->setIcon(
+            QIcon::fromTheme(
+                    "dialog-cancel",
+                    QIcon(":/icons/breeze-qownnotes/16x16/dialog-cancel.svg")));
+    button->setDefault(true);
+    ui->buttonBox->addButton(button, QDialogButtonBox::ActionRole);
 
-    connect( this->ui->buttonBox, SIGNAL( clicked( QAbstractButton* ) ), SLOT( dialogButtonClicked( QAbstractButton* ) ) );
-    connect( this, SIGNAL( finished(int) ), this, SLOT( storeSettings() ) );
+    connect(this->ui->buttonBox, SIGNAL(clicked(QAbstractButton *)),
+            SLOT(dialogButtonClicked(QAbstractButton *)));
+    connect(this, SIGNAL(finished(int)), this, SLOT(storeSettings()));
 
     QString itemName;
     QString diffHtml;
@@ -39,37 +46,39 @@ VersionDialog::VersionDialog(QScriptValue versions, MainWindow *mainWindow, QWid
     dataList = new QStringList();
 
     // init the iterator for the verions
-    QScriptValueIterator versionsIterator( versions );
+    QScriptValueIterator versionsIterator(versions);
 
     // iterate over the versions
-    while ( versionsIterator.hasNext() ) {
+    while (versionsIterator.hasNext()) {
         versionsIterator.next();
 //        qDebug() << versionsIterator.name() << ": " << versionsIterator.value().property( "timestamp" ).toString() << " - " << versionsIterator.value().property( "humanReadableTimestamp" ).toString() << " - " << versionsIterator.value().property( "diffHtml" ).toString();
 
-        itemName = versionsIterator.value().property( "humanReadableTimestamp" ).toString();
+        itemName = versionsIterator.value().property(
+                "humanReadableTimestamp").toString();
 
-        if ( itemName == "" ) {
+        if (itemName == "") {
             continue;
         }
 
-        diffHtml = versionsIterator.value().property( "diffHtml" ).toString();
-        diffHtml.replace( "<ins>", "<span style='background-color: rgb(214, 255, 199)'>" );
-        diffHtml.replace( "</ins>", "</span>" );
-        diffHtml.replace( "<del>", "<span style='background-color: rgb(255, 215, 215)'>" );
-        diffHtml.replace( "</del>", "</span>" );
-        diffHtml.replace( "\n", "<br />" );
+        diffHtml = versionsIterator.value().property("diffHtml").toString();
+        diffHtml.replace("<ins>",
+                         "<span style='background-color: rgb(214, 255, 199)'>");
+        diffHtml.replace("</ins>", "</span>");
+        diffHtml.replace("<del>",
+                         "<span style='background-color: rgb(255, 215, 215)'>");
+        diffHtml.replace("</del>", "</span>");
+        diffHtml.replace("\n", "<br />");
 
-        ui->versionListWidget->addItem( itemName );
-        diffList->append( diffHtml );
-        dataList->append( versionsIterator.value().property( "data" ).toString() );
+        ui->versionListWidget->addItem(itemName);
+        diffList->append(diffHtml);
+        dataList->append(versionsIterator.value().property("data").toString());
     }
 
-    ui->versionListWidget->setCurrentRow( 0 );
-    ui->diffBrowser->setHtml( diffList->at( 0 ) );
+    ui->versionListWidget->setCurrentRow(0);
+    ui->diffBrowser->setHtml(diffList->at(0));
 }
 
-void VersionDialog::setupMainSplitter()
-{
+void VersionDialog::setupMainSplitter() {
     this->versionSplitter = new QSplitter;
 
     this->versionSplitter->addWidget(ui->versionListWidget);
@@ -77,37 +86,34 @@ void VersionDialog::setupMainSplitter()
 
     // restore splitter sizes
     QSettings settings;
-    QByteArray state = settings.value( "versionSplitterSizes" ).toByteArray();
-    this->versionSplitter->restoreState( state );
+    QByteArray state = settings.value("versionSplitterSizes").toByteArray();
+    this->versionSplitter->restoreState(state);
 
-    this->ui->gridLayout->layout()->addWidget( this->versionSplitter );
-    this->ui->gridLayout->layout()->addWidget( ui->buttonBox );
+    this->ui->gridLayout->layout()->addWidget(this->versionSplitter);
+    this->ui->gridLayout->layout()->addWidget(ui->buttonBox);
 }
 
-void VersionDialog::storeSettings()
-{
+void VersionDialog::storeSettings() {
     // store the splitter sizes
     QSettings settings;
-    settings.setValue( "versionSplitterSizes", this->versionSplitter->saveState() );
+    settings.setValue("versionSplitterSizes",
+                      this->versionSplitter->saveState());
 }
 
-VersionDialog::~VersionDialog()
-{
+VersionDialog::~VersionDialog() {
     delete ui;
 }
 
-void VersionDialog::on_versionListWidget_currentRowChanged(int currentRow)
-{
-    ui->diffBrowser->setHtml( diffList->value( currentRow ) );
+void VersionDialog::on_versionListWidget_currentRowChanged(int currentRow) {
+    ui->diffBrowser->setHtml(diffList->value(currentRow));
 }
 
-void VersionDialog::dialogButtonClicked( QAbstractButton *button )
-{
+void VersionDialog::dialogButtonClicked(QAbstractButton *button) {
     int actionRole = button->property("ActionRole").toInt();
 
-    if ( actionRole == Restore )
-    {
-        mainWindow->setCurrentNoteText( dataList->value( ui->versionListWidget->currentRow() ) );
+    if (actionRole == Restore) {
+        mainWindow->setCurrentNoteText(
+                dataList->value(ui->versionListWidget->currentRow()));
     }
 
     this->close();