+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - implemented plural translations
   (for [Issue #65](https://github.com/pbek/QOwnNotes/issues/65))
 - updated German translation
+- added missing icons in dialogs
 
 ## 0.98
 - made a workaround for the lacking OS X theming support, icons should work